requirements.txt
wget
tgcrypto
uvloop
tornado
aiosignal
git+https://github.com/navycodes/pyrogram
pyromod==1.1.0
pytz
croniter
python-dotenv
fastapi
uvicorn
httpx[http2]==0.23.3
aiofiles
aiohttp
requests
aiorun
gunicorn==20.1.0
flask
